Resolve StackOverflowError. Source: calling StorageManager.loadUserStats() in constructor of UserStats. Move methods from Stats to UserStats and (temporarily) remove dependency of UserStats on Stats.
